Move logging into internal logger package
