+++ controllers/webhook/runcompletion_feed.go
@@ -84,7 +84,7 @@ func (rcf RunCompletionFeed) extractRunCompletionEventData(ctx context.Context, 
 }
 
 func (rcf RunCompletionFeed) fetchRunConfiguration(ctx context.Context, name *common.NamespacedName) (*pipelineshub.RunConfiguration, EventError) {
-	logger := common.LoggerFromContext(ctx)
+	logger := log.FromContext(ctx)
 	if name != nil {
 		rc := &pipelineshub.RunConfiguration{}
 		if err := rcf.client.Get(ctx, client.ObjectKey{
@@ -104,7 +104,7 @@ func (rcf RunCompletionFeed) fetchRunConfiguration(ctx context.Context, name *co
 }
 
 func (rcf RunCompletionFeed) fetchRun(ctx context.Context, name *common.NamespacedName) (*pipelineshub.Run, EventError) {
-	logger := common.LoggerFromContext(ctx)
+	logger := log.FromContext(ctx)
 	if name != nil {
 		run := &pipelineshub.Run{}
 		if err := rcf.client.Get(ctx, client.ObjectKey{
