+++ internal/log/logging.go
@@ -1,4 +1,4 @@
-package common
+package log
 
 import (
 	"context"
