+++ provider-service/base/pkg/server/provider_server.go
@@ -12,7 +12,7 @@ import (
 	"github.com/go-chi/chi/v5"
 	"github.com/go-chi/chi/v5/middleware"
 	"github.com/go-logr/logr"
-	"github.com/sky-uk/kfp-operator/pkg/common"
+	"github.com/sky-uk/kfp-operator/internal/log"
 	"github.com/sky-uk/kfp-operator/pkg/providers/base"
 	"github.com/sky-uk/kfp-operator/provider-service/base/pkg/config"
 	"github.com/sky-uk/kfp-operator/provider-service/base/pkg/server/resource"
@@ -30,7 +30,7 @@ func livenessHandler(w http.ResponseWriter, _ *http.Request) {
 }
 
 func createHandler(ctx context.Context, hr resource.HttpHandledResource) http.HandlerFunc {
-	logger := common.LoggerFromContext(ctx)
+	logger := log.LoggerFromContext(ctx)
 
 	return func(w http.ResponseWriter, r *http.Request) {
 		requestCtx := logr.NewContext(r.Context(), logger)
@@ -65,7 +65,7 @@ func createHandler(ctx context.Context, hr resource.HttpHandledResource) http.Ha
 }
 
 func updateHandler(ctx context.Context, hr resource.HttpHandledResource) http.HandlerFunc {
-	logger := common.LoggerFromContext(ctx)
+	logger := log.LoggerFromContext(ctx)
 
 	return func(w http.ResponseWriter, r *http.Request) {
 		requestCtx := logr.NewContext(r.Context(), logger)
@@ -106,7 +106,7 @@ func updateHandler(ctx context.Context, hr resource.HttpHandledResource) http.Ha
 }
 
 func deleteHandler(ctx context.Context, hr resource.HttpHandledResource) http.HandlerFunc {
-	logger := common.LoggerFromContext(ctx)
+	logger := log.LoggerFromContext(ctx)
 
 	return func(w http.ResponseWriter, r *http.Request) {
 		requestCtx := logr.NewContext(r.Context(), logger)
