+++ provider-service/base/pkg/server/resource/run.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"encoding/json"
 
-	"github.com/sky-uk/kfp-operator/pkg/common"
+	"github.com/sky-uk/kfp-operator/internal/log"
 	"github.com/sky-uk/kfp-operator/pkg/providers/base"
 )
 
@@ -17,7 +17,7 @@ func (*Run) Type() string {
 }
 
 func (r *Run) Create(ctx context.Context, body []byte) (base.Output, error) {
-	logger := common.LoggerFromContext(ctx)
+	logger := log.LoggerFromContext(ctx)
 	rd := base.RunDefinition{}
 
 	if err := json.Unmarshal(body, &rd); err != nil {
@@ -42,7 +42,7 @@ func (r *Run) Update(_ context.Context, _ string, _ []byte) (base.Output, error)
 }
 
 func (r *Run) Delete(ctx context.Context, id string) error {
-	logger := common.LoggerFromContext(ctx)
+	logger := log.LoggerFromContext(ctx)
 	if err := r.Provider.DeleteRun(ctx, id); err != nil {
 		logger.Error(err, "DeleteRun failed", "id", id)
 		return err
