+++ provider-service/base/pkg/streams/sinks/webhook_sink_unit_test.go
@@ -13,6 +13,7 @@ import (
 	"github.com/jarcoal/httpmock"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
+	"github.com/sky-uk/kfp-operator/internal/log"
 	"github.com/sky-uk/kfp-operator/pkg/common"
 	. "github.com/sky-uk/kfp-operator/provider-service/base/pkg"
 	"go.uber.org/zap/zapcore"
@@ -24,7 +25,7 @@ func TestSinksUnitSuite(t *testing.T) {
 }
 
 var _ = Context("SendEvents", func() {
-	var logger, _ = common.NewLogger(zapcore.DebugLevel)
+	var logger, _ = log.NewLogger(zapcore.DebugLevel)
 	var ctx = logr.NewContext(context.Background(), logger)
 	runCompletionEventData := common.RunCompletionEventData{}
