+++ provider-service/base/pkg/streams/sources/pub_sub_source.go
@@ -7,7 +7,7 @@ import (
 
 	"cloud.google.com/go/pubsub"
 	"github.com/go-logr/logr"
-	"github.com/sky-uk/kfp-operator/pkg/common"
+	"github.com/sky-uk/kfp-operator/internal/log"
 	. "github.com/sky-uk/kfp-operator/provider-service/base/pkg"
 )
 
@@ -33,7 +33,7 @@ func (ps *PubSubSource) ErrorOut() <-chan error {
 }
 
 func NewPubSubSource(ctx context.Context, project string, subscription string) (*PubSubSource, error) {
-	logger := common.LoggerFromContext(ctx)
+	logger := log.LoggerFromContext(ctx)
 
 	pubSubClient, err := pubsub.NewClient(ctx, project)
 	if err != nil {
