+++ provider-service/vai/internal/provider/provider.go
@@ -7,6 +7,7 @@ import (
 
 	aiplatform "cloud.google.com/go/aiplatform/apiv1"
 	"cloud.google.com/go/aiplatform/apiv1/aiplatformpb"
+	"github.com/sky-uk/kfp-operator/internal/log"
 	"github.com/sky-uk/kfp-operator/pkg/common"
 	"github.com/sky-uk/kfp-operator/pkg/providers/base"
 	"github.com/sky-uk/kfp-operator/provider-service/base/pkg/server/resource"
@@ -121,7 +122,7 @@ func (vaip *VAIProvider) DeletePipeline(ctx context.Context, id string) error {
 }
 
 func (vaip *VAIProvider) CreateRun(ctx context.Context, rd base.RunDefinition) (string, error) {
-	logger := common.LoggerFromContext(ctx)
+	logger := log.LoggerFromContext(ctx)
 
 	pipelinePath, err := util.PipelineStorageObject(
 		rd.PipelineName,
@@ -176,7 +177,7 @@ func (vaip *VAIProvider) CreateRunSchedule(
 	ctx context.Context,
 	rsd base.RunScheduleDefinition,
 ) (string, error) {
-	logger := common.LoggerFromContext(ctx)
+	logger := log.LoggerFromContext(ctx)
 
 	pipelinePath, err := util.PipelineStorageObject(
 		rsd.PipelineName,
@@ -236,7 +237,7 @@ func (vaip *VAIProvider) UpdateRunSchedule(
 	rsd base.RunScheduleDefinition,
 	_ string,
 ) (string, error) {
-	logger := common.LoggerFromContext(ctx)
+	logger := log.LoggerFromContext(ctx)
 
 	pipelinePath, err := util.PipelineStorageObject(
 		rsd.PipelineName,
