+++ provider-service/vai/internal/runcompletion/event_flow.go
@@ -7,6 +7,7 @@ import (
 
 	aiplatform "cloud.google.com/go/aiplatform/apiv1"
 	"cloud.google.com/go/aiplatform/apiv1/aiplatformpb"
+	"github.com/sky-uk/kfp-operator/internal/log"
 	"github.com/sky-uk/kfp-operator/pkg/common"
 	. "github.com/sky-uk/kfp-operator/provider-service/base/pkg"
 	"github.com/sky-uk/kfp-operator/provider-service/base/pkg/streams"
@@ -82,7 +83,7 @@ func NewEventFlow(config *config.VAIProviderConfig, pipelineJobClient *aiplatfor
 
 func (vef *EventFlow) Start(ctx context.Context) {
 	go func() {
-		logger := common.LoggerFromContext(ctx)
+		logger := log.LoggerFromContext(ctx)
 		for msg := range vef.in {
 			logger.Info("in VAI flow - received message", "message", msg.Message)
 			runCompletionEventData, err := vef.runCompletionEventDataForRun(ctx, msg.Message)
@@ -111,7 +112,7 @@ func (vef *EventFlow) runCompletionEventDataForRun(ctx context.Context, runId st
 		Name: vef.ProviderConfig.PipelineJobName(runId),
 	})
 	if err != nil {
-		common.LoggerFromContext(ctx).Error(err, "failed to fetch pipeline job")
+		log.LoggerFromContext(ctx).Error(err, "failed to fetch pipeline job")
 		return nil, err
 	}
 	return vef.toRunCompletionEventData(ctx, job, runId)
@@ -195,7 +196,7 @@ func (vef *EventFlow) toRunCompletionEventData(ctx context.Context, job *aiplatf
 
 	if !completed {
 		err := errors.New(PipelineJobNotFinishedErr)
-		common.LoggerFromContext(ctx).Error(err, "run-id", runId)
+		log.LoggerFromContext(ctx).Error(err, "run-id", runId)
 		return nil, err
 	}
